Allow users to control the name of the base VM
When an Installer App is provided, the current behavior is to create
a new VM named `anka-base-XXXXXXXXXX`. The trailing X's are random
characters and in a CI system responsible for building images, this
creates cruft.

This change introduces new behavior when a user specifies both a
`source_vm_name` and an `installer_app`. If both are specified, rather
than creating an image named `anka-base-XXXXXXXXXX`, it will create
the VM using the `source_vm_name`. Then, on subsequent rebuilds, the
plugin will use the pre-existing `source_vm_name` rather than
re-doing the install procedure.

Signed-off-by: Tom Duffield <tom@chef.io>